fix results object filename
